requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-postgres
langgraph
pgvector
psycopg2-binary
pypdf
pandas
python-multipart
python-dotenv
openai
lark
